Convert button directive to component
